Refactor com_atproto::repo
